+++ source/extensions/lightspeed.trex.mesh_properties.shared.widget/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [1.0.3] - 2023-10-16
+### Fixed
+- Proper filter lights for transform
+
 ## [1.0.2] - 2023-09-20
 ### Added
 - Support doubleSided USD mesh attribute in UI