+++ source/extensions/lightspeed.trex.properties_pane.shared.mod_setup.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0] - 2023-10-16
+### Added
+- Added tests + fix selected item not centered in the scroll frame
+
 ## [1.0.0] - 2023-05-18
 ### Added
 - Created